Moved the r(z) and z(r) functionality into its own file called cosmology.c.
