+++ cosmology.h
@@ -2,10 +2,13 @@
   This file contains the assumed cosmology we will use.
   It is in a separated file for easy access.
 
-  TODO: This header file contains the prototypes for
-  TODO: cosmology.c
-  TODO: See accompanying .c file for more information.
+  This header file contains the prototypes for
+  cosmology.c
+  See accompanying .c file for more information.
 */
+
+#include <stdio.h>
+#include <math.h>
 
 #ifndef COSMOLOGY_H
 
@@ -14,8 +17,9 @@
 #define cosmo_lambda 0.7
 #define cosmo_h 0.7
 
-/*
-  
-*/
+double comovingDistance(double z);
+void setupRedshiftInterp(double* *r, double* *z, int* numInterpPts);
+double interpRedshift(double rQuery, double* r, double* z, int numInterpPts);
+double interpDist(double zQuery, double* r, double* z, int numInterpPts);
 
 #endif
